requirements.txt
fastapi
uvicorn
murf
jinja2
aiofiles
python-multipart
python-dotenv
requests
